Add files via upload
all this file is used in the paper "Conformal prediction for functional Ordinary kriging". 
This paper is uploded on arXiv with doi https://doi.org/10.48550/arXiv.2409.20084.
File "function.R" is the file with all function used.
File "simulation.R" is the file with the simuletion study.
In the file "performance_index.R" we calculate indices to evaluete the performance of conformal prediction.
In the file "example.R" we make an example of conformal prediction used with ordinary kriging on real data set.